Update README and add comments to CustomerRepositoryImpl
The README file's docker command has been updated to reflect the change in the MySQL password. Also, comments have been added to the insert method in the CustomerRepositoryImpl file for better understanding of the code functionality. Changes in the .idea/workspace.xml aim to enhance code organization and update commit messages.